Add ability to batch commands to the board
Given the multi threaded concurrent nature of Elixir, is possible that 2 LEDControllers would try to execute the send_pixels methods at the same time resulting in interleaved commands and weird behavior.

Here we can send a batch of commands and since the board will hand a batch completely before moving to the next message, we are guarenteed that a batch is not interuppted.